refactor: move scraping functs to helper_functs
